requirements.txt
# Core dependencies
celery[redis]==5.3.4
redis>=4.5.2,<5.0.0
kombu==5.3.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-timeout==2.2.0
pytest-xdist==3.5.0  # 並行測試
pytest-html==4.1.1   # HTML 報告

# Monitoring
flower==2.0.1
prometheus-client==0.19.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
colorlog==6.8.0
python-json-logger==2.0.7
tabulate==0.9.0

# Development
ipython==8.18.1
ipdb==0.13.13

# Type checking
celery-types==0.20.0
mypy==1.7.1